pyproject.toml
[tool.poetry]
name = "zenn-cloudfront-cognito-apigw"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
boto3 = "^1.24.17"
PyJWT = "^2.4.0"


[tool.poetry.dev-dependencies]
jupyterlab = "^3.2.9"
isort = "^5.10.1"
mypy = "^0.931"
flake8 = "^4.0.1"
pyproject-flake8 = "^0.0.1-alpha.2"
flake8-builtins = "^1.5.3"
pep8-naming = "^0.12.1"
flake8-cognitive-complexity = "^0.1.0"
black = "^22.1.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
target-version = ['py38', 'py39']
line-length = 120
exclude = """
/(
    .git
    | .github
    | .mypy_cache
    | .pytest_cache
    | .venv
    | cdkv2
    | data
    | dockerfiles
    | docs
    | htmlcov
    | frontend
    | notebook
/)
"""

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
profile = "black"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

# options: https://flake8.pycqa.org/en/latest/user/options.html
[tool.flake8]
max-line-length = 120
max-complexity = 20
ignore = "E203,E266,W503"
per-file-ignores = "*/__init__.py:F401 */log.py:N802"
max-cognitive-complexity = 20
